use TimeEntry start date as date which it belongs to
a TimeEntry can start at 22:00 and end at the next day 02:00 for instance.
Or if someone forgets to stop the TimeClock manually it could cover multiple days (weekend).

with this patch, the time entry overview and the report view uses the startDate as date which it belongs to.
e.g. a TimeEntry starting Friday 08:00 and ending Monday 08:00 (duration 3 days) belongs to Friday with a workDuration of 72 hours.