fix missing query in users settings links
given user searches for "klaus"
then user selects "klaus"
then user selects "Berechtigungen"

expected: search input still contains the query

so the user can click the x on the query input to delete the query and to see all persons in the list

issue with this solution:

given user deletes the query
given user clicks "Arbeitszeit"
then the search input is filled with the query again.

however, I think this behaviour is ok...
we would have to update the links otherwise.
this is doable... but I don't have the muse right now :-)