WIP: do not loose user input rendering settings page
I don't think we should use the Optional field in the domain?
just keep the random -1 value, maybe?

to not loose the user input on rendering the settings page again
after a preview request, we just have to keep the value in the
dto. don't map an invalid value like 'asd' to empty field.

closes #1254